requirements.txt
fastapi==0.115.12
python-multipart==0.0.20
pydantic==2.10.2
pydantic_settings==2.6.1
langchain==0.3.25
langchain-community==0.3.24
langchain-openai==0.3.22
pymupdf==1.26.1
pillow==11.3.0
google-cloud-vision==3.10.2